sound: rename sound_queue to queue
