+++ sound/libsoundsharedringbuffer/sound_queue.c
@@ -1,4 +1,4 @@
-#include <sddf/sound/sound_queue.h>
+#include <sddf/sound/queue.h>
 #include <sddf/util/fence.h>
 
 void sddf_snd_ring_init(sddf_snd_ring_state_t *ring_state, uint32_t buffer_count)
